Add git_find_version_tag() util
Function is able to find the nearest git tag that matches a default glob pattern. It is not 100% accurate because the "--match" does not accept advanced regex patterns - it only supports glob patterns. In any case, provided that a project adheres to semver, then this should be able to get the job done.