Fix executor path is invalid when package is consumed
For some strange reason, the "does executor exist" check fails, when this package is consumed by another project. By removing the check, it seems that adding ctests using the executor works... Not sure why this fails in the first place

the default path seems to fail when this package is consumed by another project.